Created a TIME package and setup.py
