+++ TIME/__init__.py
@@ -1,0 +1,9 @@
+"""
+This package contains an implementation of TIME, an algorithm used to combine
+macroscopic tractography information with microscopic multi-fixel model
+estimates in order to improve the accuracy in the estimation of the
+microstructural properties of neural fibers in a specified tract.
+"""
+
+__version__ = "0.0.1a0"
+__author__ = 'Nicolas Delinte'
